2/requirements.txt
# Black2 X402 Integration
# Web Framework
fastapi>=0.115.0
uvicorn>=0.30.6
python-multipart>=0.0.9

# Cryptography
cryptography>=43.0.3
PyNaCl>=1.5.0

# Database
aiosqlite>=0.20.0

# HTTP Client
httpx>=0.27.2

# Utilities
python-dotenv>=1.0.1
pydantic>=2.9.2

# X402 Payment Bridge
uvd-x402-sdk>=0.1.0

# Testing
pytest>=8.3.4
pytest-asyncio>=0.25.2